[scanner] refactor: split api.ts into domain-aligned modules
Splits the monolithic api.ts (987 lines) into domain-specific modules:
- api/cluster.ts — Cluster and namespace operations
- api/settings.ts — Settings and configuration operations
- api/dashboard.ts — Dashboard and card management
- api/agent.ts — Agent and MCP operations
- api/client.ts — Core HTTP client class
- api/utils.ts — Helper functions (safeJson, authFetch)
- api/index.ts — Barrel export for backward compatibility

The main api.ts now re-exports from ./api/index for full backward compatibility.

Fixes #19013
