+++ web/src/lib/api.ts
@@ -1,987 +1,14 @@
-/* eslint-disable max-lines -- TODO: split this file (tracked by #15790) */
-import { MCP_HOOK_TIMEOUT_MS, BACKEND_HEALTH_CHECK_TIMEOUT_MS, STORAGE_KEY_USER_CACHE, STORAGE_KEY_HAS_SESSION, DEMO_TOKEN_VALUE, FETCH_DEFAULT_TIMEOUT_MS } from './constants'
-import { clearStoredAuthToken, getStoredAuthToken, getStoredAuthTokenSync } from './authToken'
-import { emitSessionExpired, emitHttpError } from './analytics'
-import {
-  reportBackendAvailable,
-  reportBackendUnavailable,
-  shouldMarkBackendUnavailable,
-} from './backendHealthEvents'
-import { reportAppError } from './errors/handleError'
-
-const API_BASE = ''
-const DEFAULT_TIMEOUT = MCP_HOOK_TIMEOUT_MS
-const BACKEND_CHECK_INTERVAL = 10_000 // 10 seconds between backend checks when unavailable
-/** How long to trust a cached backend-availability check (5 minutes) */
-const BACKEND_CACHE_TTL_MS = 300_000
 /**
- * Routes whose 5xx responses indicate an optional upstream/dependency issue,
- * not that the console backend itself is down.
- */
-const BACKEND_OUTAGE_EXEMPT_PREFIXES = ['/api/kagent/', '/api/kagenti-provider/']
-/** Delay before redirecting to login after session expiry (lets user see the banner) */
-const SESSION_EXPIRY_REDIRECT_MS = 3_000
-const TOKEN_REFRESH_HEADER = 'X-Token-Refresh' // server signals when token should be refreshed
-/** Endpoint used to invalidate the HttpOnly auth cookie on the server side (#6061). */
-const AUTH_LOGOUT_ENDPOINT = '/auth/logout'
-
-// Public API paths that don't require authentication (served without JWT on the backend)
-const PUBLIC_API_PREFIXES = ['/api/missions/browse', '/api/missions/file', '/api/compliance/']
-
-// Error class for unauthenticated requests
-export class UnauthenticatedError extends Error {
-  constructor() {
-    super('No authentication token available')
-    this.name = 'UnauthenticatedError'
-  }
-}
-
-// Error class for 401 unauthorized responses (invalid/expired token)
-export class UnauthorizedError extends Error {
-  constructor() {
-    super('Token is invalid or expired')
-    this.name = 'UnauthorizedError'
-  }
-}
-
-/** localStorage key for global API rate-limit backoff deadline (epoch ms). */
-const STORAGE_KEY_RATE_LIMIT_UNTIL = 'kc-api-rate-limit-until'
-/** Default Retry-After when the header is missing or unparseable. */
-const DEFAULT_RATE_LIMIT_RETRY_AFTER_S = 60
-
-export class RateLimitError extends Error {
-  retryAfter: number
-  constructor(retryAfter: number) {
-    super(`Rate limited. Try again in ${retryAfter} seconds.`)
-    this.name = 'RateLimitError'
-    this.retryAfter = retryAfter
-  }
-}
-
-function handle429(response: Response): never {
-  const retryAfterRaw = response.headers.get('Retry-After')
-  const retryAfter = retryAfterRaw ? parseInt(retryAfterRaw, 10) : DEFAULT_RATE_LIMIT_RETRY_AFTER_S
-  const effectiveRetry = Number.isFinite(retryAfter) && retryAfter > 0
-    ? retryAfter : DEFAULT_RATE_LIMIT_RETRY_AFTER_S
-  try {
-    localStorage.setItem(STORAGE_KEY_RATE_LIMIT_UNTIL,
-      String(Date.now() + effectiveRetry * 1000))
-  } catch (error: unknown) {
-    reportAppError(error, {
-      context: '[API] Failed to persist rate-limit retry window',
-      level: 'warn',
-      fallbackMessage: 'rate limit storage write failed',
-    })
-  }
-  throw new RateLimitError(effectiveRetry)
-}
-
-// Debounce 401 handling to avoid multiple simultaneous logouts
-let handling401 = false
-/** Safety cap: reset the 401 debounce flag after this many ms so future
- *  auth failures aren't permanently silenced if the redirect doesn't fire (#3899). */
-const HANDLING_401_RESET_MS = 10_000
-/** Short timeout on the session-verify probe so a hung backend doesn't block
- *  the session-expired UI indefinitely (#8372). */
-const SESSION_VERIFY_TIMEOUT_MS = 3_000
-/** Endpoint used to verify the HttpOnly cookie is still valid. A 200 here
- *  means the cookie still authenticates, so a 401 from another endpoint was
- *  endpoint-specific (not a session expiry). */
-const AUTH_VERIFY_ENDPOINT = '/api/me'
-
-/**
- * Handle 401 Unauthorized responses by clearing auth state and redirecting to login.
- * This is debounced to avoid multiple simultaneous logouts from parallel API calls.
- * The flag auto-resets after HANDLING_401_RESET_MS so a failed redirect doesn't
- * permanently block all API calls.
+ * api.ts — Main API client (now a barrel re-export).
+ * Split into domain-specific modules per issue #19013:
+ * - api/client.ts — Core HTTP client class
+ * - api/utils.ts — Helper functions (safeJson, authFetch)
+ * - api/cluster.ts — Cluster and namespace operations
+ * - api/settings.ts — Settings and configuration operations
+ * - api/dashboard.ts — Dashboard and card management
+ * - api/agent.ts — Agent and MCP operations
+ * - api/index.ts — Barrel export
  *
- * Before nuking the session we re-verify via /api/me. If the cookie-based
- * session is still valid, the 401 was specific to the originating endpoint
- * (e.g. a route that requires elevated scope, or a backend race) and we must
- * NOT show "Session expired" nor redirect — doing so would bounce the user
- * through /login?reason=session_expired and leave a stale `?reason=…` query
- * param on the landing page (#8372).
+ * This file maintained for backward compatibility.
  */
-function handle401(): void {
-  if (handling401) return
-  handling401 = true
-
-  // Auto-reset the flag after a safety timeout so the app isn't permanently
-  // blocked if the redirect fails (e.g. service-worker intercept, popup blocker).
-  setTimeout(() => {
-    handling401 = false
-  }, HANDLING_401_RESET_MS)
-
-  // Verify the cookie-based session before treating this as an expiry. If
-  // /api/me comes back 200, the session is still valid — abort the
-  // session-expired flow entirely (#8372).
-  fetch(`${API_BASE}${AUTH_VERIFY_ENDPOINT}`, {
-    credentials: 'include',
-    signal: AbortSignal.timeout(SESSION_VERIFY_TIMEOUT_MS),
-  }).then(verifyResponse => {
-    if (verifyResponse.ok) {
-      reportAppError(new Error('[API] 401 received but /api/me still 200 — endpoint-specific failure, keeping session'), {
-        context: '[API]',
-        level: 'warn',
-      })
-      handling401 = false
-      return
-    }
-    if (verifyResponse.status === 429) {
-      reportAppError(new Error('[API] 401 received but /api/me returned 429 (rate-limited) — keeping session'), {
-        context: '[API]',
-        level: 'warn',
-      })
-      handling401 = false
-      return
-    }
-    performSessionExpiry()
-  }).catch((error: unknown) => {
-    reportAppError(error, {
-      context: '[API] Session verify probe failed during 401 handling',
-      level: 'warn',
-      fallbackMessage: 'session verify probe failed',
-    })
-    // Verify probe failed (network error / timeout / no backend). Treat the
-    // 401 as authoritative and run the normal expiry flow.
-    performSessionExpiry()
-  })
-}
-
-/** Second half of handle401: banner + cookie invalidation + redirect. */
-function performSessionExpiry(): void {
-  reportAppError(new Error('[API] Received 401 Unauthorized - token invalid or expired, logging out'), {
-    context: '[API]',
-    level: 'warn',
-  })
-
-  // Show an in-page notification before redirecting (DOM-injected, no React dependency)
-  showSessionExpiredBanner()
-
-  emitSessionExpired()
-
-  // Fire-and-forget: invalidate the HttpOnly auth cookie on the server so that
-  // a stale cookie doesn't resurrect the session on the next page load (#6061).
-  // We pass credentials:'include' so the browser sends the cookie. We don't
-  // await the response and we ignore failures — the client-side clear below
-  // is the source of truth for logout.
-  const expiredToken = getStoredAuthTokenSync()
-  try {
-    const headers: Record<string, string> = { 'Content-Type': 'application/json' }
-    if (expiredToken && expiredToken !== DEMO_TOKEN_VALUE) {
-      headers['Authorization'] = `Bearer ${expiredToken}`
-    }
-    fetch(`${API_BASE}${AUTH_LOGOUT_ENDPOINT}`, {
-      method: 'POST',
-      headers,
-      credentials: 'include',
-      signal: AbortSignal.timeout(FETCH_DEFAULT_TIMEOUT_MS),
-    }).catch((error: unknown) => {
-      reportAppError(error, {
-        context: '[API] Logout request failed during session expiry',
-        level: 'warn',
-        fallbackMessage: 'logout request failed',
-      })
-      // Backend unreachable — cookie will expire naturally
-    })
-  } catch (error: unknown) {
-    reportAppError(error, {
-      context: '[API] Failed to initiate logout request during session expiry',
-      level: 'warn',
-      fallbackMessage: 'logout request setup failed',
-    })
-    // fetch() threw synchronously (very rare) — ignore
-  }
-
-  // Clear auth state
-  clearStoredAuthToken()
-  localStorage.removeItem(STORAGE_KEY_USER_CACHE)
-  localStorage.removeItem(STORAGE_KEY_HAS_SESSION)
-
-  // Redirect to login after a delay so the user sees the banner
-  setTimeout(() => {
-    window.location.href = '/login?reason=session_expired'
-  }, SESSION_EXPIRY_REDIRECT_MS)
-}
-
-/**
- * Inject a DOM-based notification banner for session expiry.
- * This runs outside React so it works from any context (API client, background fetches, etc).
- */
-function showSessionExpiredBanner(): void {
-  // Avoid duplicates
-  if (document.getElementById('session-expired-banner')) return
-
-  const toast = document.createElement('div')
-  toast.id = 'session-expired-banner'
-  toast.style.cssText = `
-    position: fixed; bottom: 24px; left: 50%; transform: translateX(-50%); z-index: 99999;
-    display: flex; align-items: center; gap: 12px;
-    padding: 12px 20px;
-    background: rgba(234,179,8,0.15);
-    border: 1px solid rgba(234,179,8,0.4);
-    border-radius: 8px; backdrop-filter: blur(8px);
-    color: #fbbf24; font-family: system-ui, sans-serif; font-size: 14px;
-    animation: slideUp 0.3s ease-out;
-  `
-  toast.innerHTML = `
-    <svg xmlns="http://www.w3.org/2000/svg" width="18" height="18" viewBox="0 0 24 24" fill="none"
-      stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round">
-      <path d="m21.73 18-8-14a2 2 0 0 0-3.48 0l-8 14A2 2 0 0 0 4 21h16a2 2 0 0 0 1.73-3"/>
-      <path d="M12 9v4"/><path d="M12 17h.01"/>
-    </svg>
-    <span><strong>Session expired</strong> — Redirecting to sign in...</span>
-  `
-
-  // Reuse a single <style> element to avoid unbounded DOM growth
-  const STYLE_ID = 'session-banner-animation'
-  if (!document.getElementById(STYLE_ID)) {
-    const style = document.createElement('style')
-    style.id = STYLE_ID
-    style.textContent = `@keyframes slideUp { from { transform: translateX(-50%) translateY(100%); opacity: 0; } to { transform: translateX(-50%) translateY(0); opacity: 1; } }`
-    document.head.appendChild(style)
-  }
-  document.body.appendChild(toast)
-}
-
-// Error class for backend unavailable
-export class BackendUnavailableError extends Error {
-  constructor() {
-    super('Backend API is currently unavailable')
-    this.name = 'BackendUnavailableError'
-  }
-}
-
-// Backend availability tracking with localStorage persistence
-const BACKEND_STATUS_KEY = 'kc-backend-status'
-let backendLastCheckTime = 0
-let backendAvailable: boolean | null = null // null = unknown, true = available, false = unavailable
-let backendCheckPromise: Promise<boolean> | null = null
-
-// Initialize from localStorage
-try {
-  const stored = localStorage.getItem(BACKEND_STATUS_KEY)
-  if (stored) {
-    const { available, timestamp } = JSON.parse(stored)
-    // Use cached status if checked within the last 5 minutes
-    if (Date.now() - timestamp < BACKEND_CACHE_TTL_MS) {
-      backendAvailable = available
-      backendLastCheckTime = timestamp
-    }
-  }
-} catch (error: unknown) {
-  reportAppError(error, {
-    context: '[api] failed to load cached backend status',
-    level: 'warn',
-    fallbackMessage: 'backend status cache read failed',
-  })
-}
-
-/**
- * Check backend availability - only makes ONE request, all others wait
- * Caches result in localStorage to avoid repeated checks across page loads
- * @param forceCheck - If true, ignores cache and always checks (used by login)
- */
-export async function checkBackendAvailability(forceCheck = false): Promise<boolean> {
-  // If we already know the status and it was checked recently, return it.
-  // The TTL gate must always run so a previously-available backend is
-  // re-probed periodically instead of being cached forever.
-  if (!forceCheck && backendAvailable !== null) {
-    const now = Date.now()
-    if (now - backendLastCheckTime < BACKEND_CHECK_INTERVAL) {
-      return backendAvailable
-    }
-  }
-
-  // If a check is already in progress, wait for it
-  if (backendCheckPromise) {
-    return backendCheckPromise
-  }
-
-  // Start a new check
-  backendCheckPromise = (async () => {
-    try {
-      const response = await fetch(`${API_BASE}/health`, {
-        method: 'GET',
-        signal: AbortSignal.timeout(BACKEND_HEALTH_CHECK_TIMEOUT_MS),
-      })
-      // Backend is available if it responds at all (even non-200)
-      // Only 5xx or network errors indicate backend is down
-      backendAvailable = response.status < 500
-      backendLastCheckTime = Date.now()
-      // Cache to localStorage
-      try {
-        localStorage.setItem(BACKEND_STATUS_KEY, JSON.stringify({
-          available: backendAvailable,
-          timestamp: backendLastCheckTime,
-        }))
-      } catch (error: unknown) {
-        reportAppError(error, {
-          context: '[api] failed to cache backend status',
-          fallbackMessage: 'backend status cache write failed',
-        })
-      }
-      return backendAvailable
-    } catch (error: unknown) {
-      reportAppError(error, {
-        context: '[api] backend availability check failed',
-        level: 'warn',
-        fallbackMessage: 'backend availability check failed',
-      })
-      backendAvailable = false
-      backendLastCheckTime = Date.now()
-      // Only cache failures in memory — do NOT persist false to localStorage.
-      // Persisting false causes the stuck state where a fresh page load inherits
-      // a stale "backend down" flag and blocks all API calls indefinitely.
-      return false
-    } finally {
-      backendCheckPromise = null
-    }
-  })()
-
-  return backendCheckPromise
-}
-
-/** #6055 — number of retry attempts for checkOAuthConfiguredWithRetry during backend startup. */
-const OAUTH_STARTUP_RETRY_ATTEMPTS = 5
-/** #6055 — delay (ms) between retry attempts. */
-const OAUTH_STARTUP_RETRY_DELAY_MS = 2_000
-
-/**
- * #6055 — Retry wrapper around checkOAuthConfigured() for bootstrap races
- * where the frontend loads before the backend is accepting connections.
- * Retries up to OAUTH_STARTUP_RETRY_ATTEMPTS times, sleeping
- * OAUTH_STARTUP_RETRY_DELAY_MS between attempts, exiting early as soon as
- * the backend comes up.
- */
-export async function checkOAuthConfiguredWithRetry(): Promise<{ backendUp: boolean; oauthConfigured: boolean }> {
-  let lastResult: { backendUp: boolean; oauthConfigured: boolean } = { backendUp: false, oauthConfigured: false }
-  for (let attempt = 0; attempt < OAUTH_STARTUP_RETRY_ATTEMPTS; attempt++) {
-    try {
-      const result = await checkOAuthConfigured()
-      lastResult = result
-      if (result.backendUp) return result
-    } catch (error: unknown) {
-      reportAppError(error, {
-        context: '[api] OAuth startup check failed; retrying',
-        level: 'warn',
-        fallbackMessage: 'oauth startup check failed',
-      })
-    }
-    if (attempt < OAUTH_STARTUP_RETRY_ATTEMPTS - 1) {
-      await new Promise((resolve) => setTimeout(resolve, OAUTH_STARTUP_RETRY_DELAY_MS))
-    }
-  }
-  return lastResult
-}
-
-/**
- * Check if the backend has OAuth configured by reading the /health endpoint.
- * Returns { backendUp, oauthConfigured }.
- */
-export async function checkOAuthConfigured(): Promise<{ backendUp: boolean; oauthConfigured: boolean }> {
-  try {
-    const response = await fetch(`${API_BASE}/health`, {
-      method: 'GET',
-      signal: AbortSignal.timeout(BACKEND_HEALTH_CHECK_TIMEOUT_MS),
-    })
-    if (!response.ok) return { backendUp: false, oauthConfigured: false }
-    // Parse JSON through the shared helper to avoid unhandled-rejection races in
-    // Firefox and keep fallback behavior consistent.
-    const data = await safeParseJsonOrNull<{ oauth_configured?: boolean }>(
-      response,
-      '[api] /health OAuth config parse failed',
-    )
-    if (!data) return { backendUp: false, oauthConfigured: false }
-    return {
-      // Any successful /health response means the backend is reachable.
-      // A "degraded" status (e.g. all clusters unreachable) should NOT
-      // flip the app into demo mode — only a network failure should (#5401).
-      backendUp: true,
-      oauthConfigured: !!data.oauth_configured,
-    }
-  } catch (error: unknown) {
-    reportAppError(error, {
-      context: '[api] OAuth configured check failed',
-      level: 'warn',
-      fallbackMessage: 'oauth configured check failed',
-    })
-    return { backendUp: false, oauthConfigured: false }
-  }
-}
-
-function markBackendFailure(status?: number): void {
-  backendAvailable = false
-  backendLastCheckTime = Date.now()
-  reportBackendUnavailable('http', status)
-  // Don't persist false to localStorage — only keep in memory.
-  // Persisting false causes fresh page loads to inherit stale "backend down" state.
-  try {
-    localStorage.removeItem(BACKEND_STATUS_KEY)
-  } catch (error: unknown) {
-    reportAppError(error, {
-      context: '[api] failed to clear backend status cache',
-      fallbackMessage: 'backend status cache clear failed',
-    })
-  }
-}
-
-function markBackendSuccess(status?: number): void {
-  backendAvailable = true
-  backendLastCheckTime = Date.now()
-  reportBackendAvailable('http', status)
-  try {
-    localStorage.setItem(BACKEND_STATUS_KEY, JSON.stringify({
-      available: true,
-      timestamp: backendLastCheckTime,
-    }))
-  } catch (error: unknown) {
-    reportAppError(error, {
-      context: '[api] failed to cache backend success',
-      fallbackMessage: 'backend success cache write failed',
-    })
-  }
-}
-
-function createErrorWithCause(message: string, cause: unknown): Error {
-  const error = new Error(message) as Error & { cause?: unknown }
-  error.cause = cause
-  return error
-}
-
-async function safeReadTextOrEmpty(response: Response, context: string): Promise<string> {
-  try {
-    return await response.text()
-  } catch (error: unknown) {
-    reportAppError(error, {
-      context,
-      level: 'warn',
-      fallbackMessage: 'failed to read response text',
-    })
-    return ''
-  }
-}
-
-async function safeParseJsonOrNull<T = unknown>(response: Response, context: string): Promise<T | null> {
-  const contentLength = response.headers.get('content-length')
-
-  if (response.status === 204 || contentLength === '0') {
-    return {} as T
-  }
-
-  try {
-    const text = await response.text()
-    if (!text || text.trim() === '') {
-      return {} as T
-    }
-    return JSON.parse(text) as T
-  } catch (error: unknown) {
-    reportAppError(error, {
-      context,
-      level: 'warn',
-      fallbackMessage: 'failed to parse response JSON',
-    })
-    return null
-  }
-}
-
-/**
- * Check if the backend is known to be unavailable.
- * Returns true if backend is definitely unavailable (checked recently and failed).
- * Returns false if backend is available or status is unknown.
- */
-export function isBackendUnavailable(): boolean {
-  if (backendAvailable === null) return false // Unknown - allow first request
-  if (backendAvailable) return false // Available
-
-  // Check if enough time has passed for a recheck
-  const now = Date.now()
-  if (now - backendLastCheckTime >= BACKEND_CHECK_INTERVAL) {
-    return false // Allow a recheck
-  }
-
-  return true // Known unavailable
-}
-
-class ApiClient {
-  private refreshInProgress: Promise<void> | null = null
-
-  /**
-   * Silently refresh the JWT token in the background.
-   * Called when the server returns X-Token-Refresh header indicating the token
-   * has passed 50% of its lifetime and should be renewed.
-   */
-  private silentRefresh(): void {
-    if (this.refreshInProgress) return
-    this.refreshInProgress = (async () => {
-      try {
-        // #8108 — /auth/refresh must NOT receive the Authorization header.
-        // Backend RefreshToken revokes the JTI of whatever bearer is presented
-        // before minting the replacement. Sending the stale localStorage token
-        // would revoke a token we still rely on for the rest of the session
-        // and race against the cookie delivery. Cookie-only flow: send the
-        // HttpOnly kc_auth cookie via credentials + CSRF header only.
-        const response = await fetch(`${API_BASE}/auth/refresh`, {
-          method: 'POST',
-          headers: {
-            'Content-Type': 'application/json',
-            // #6588 — CSRF gate on /auth/refresh
-            'X-Requested-With': 'XMLHttpRequest',
-          },
-          credentials: 'same-origin',
-          signal: AbortSignal.timeout(FETCH_DEFAULT_TIMEOUT_MS),
-        })
-        if (response.ok) {
-          // #6590 — /auth/refresh delivers the new JWT exclusively via the
-          // HttpOnly kc_auth cookie; the JSON body carries only
-          // { refreshed: true, onboarded }. There is no token to copy into
-          // localStorage. The browser sends the refreshed cookie automatically
-          // on subsequent requests, and the JWTAuth middleware reads it.
-          // Nothing else to do here on success.
-          try {
-            localStorage.setItem(STORAGE_KEY_HAS_SESSION, 'true')
-          } catch (error: unknown) {
-            reportAppError(error, {
-              context: '[api] failed to cache session marker after refresh',
-              level: 'warn',
-              fallbackMessage: 'session marker cache write failed',
-            })
-          }
-        }
-      } catch (error: unknown) {
-        reportAppError(error, {
-          context: '[api] silent refresh failed',
-          level: 'warn',
-          fallbackMessage: 'token refresh failed',
-        })
-        // Silent refresh failure is non-fatal — the current token is still valid
-      } finally {
-        this.refreshInProgress = null
-      }
-    })()
-  }
-
-  /**
-   * Check the response for the X-Token-Refresh header and trigger a
-   * background refresh if present.
-   */
-  private checkTokenRefresh(response: Response): void {
-    if (response.headers.get(TOKEN_REFRESH_HEADER) === 'true') {
-      this.silentRefresh()
-    }
-  }
-
-  private async getHeaders(): Promise<Record<string, string>> {
-    const headers: Record<string, string> = {
-      'Content-Type': 'application/json',
-      // #8830 — the /api group middleware rejects state-changing requests
-      // (POST/PUT/DELETE/PATCH) without this header. Harmless on GET.
-      'X-Requested-With': 'XMLHttpRequest',
-    }
-    const token = await getStoredAuthToken()
-    if (token) {
-      headers['Authorization'] = `Bearer ${token}`
-    }
-    return headers
-  }
-
-  private async hasToken(): Promise<boolean> {
-    const token = await getStoredAuthToken()
-    if (token && token !== DEMO_TOKEN_VALUE) return true
-    // #6590 / #8087 — A cookie-only session has no JS-readable token, only
-    // the HttpOnly kc_auth cookie. The kc-has-session marker is set after
-    // /auth/refresh succeeds; treat its presence as a real session so API
-    // calls go through (the cookie is sent automatically same-origin and
-    // JWTAuth middleware accepts it).
-    try {
-      return localStorage.getItem(STORAGE_KEY_HAS_SESSION) === 'true'
-    } catch (error: unknown) {
-      reportAppError(error, {
-        context: '[api] failed to read session marker',
-        level: 'warn',
-        fallbackMessage: 'session marker read failed',
-      })
-      return false
-    }
-  }
-
-  private createAbortController(timeout: number): { controller: AbortController; timeoutId: ReturnType<typeof setTimeout> } {
-    const controller = new AbortController()
-    const timeoutId = setTimeout(() => controller.abort(), timeout)
-    return { controller, timeoutId }
-  }
-
-  async get<T = unknown>(path: string, options?: { headers?: Record<string, string>; timeout?: number; requiresAuth?: boolean; signal?: AbortSignal }): Promise<{ data: T }> {
-    // Skip API calls to protected endpoints when not authenticated
-    const isPublicPath = PUBLIC_API_PREFIXES.some(prefix => path.startsWith(prefix))
-    if (options?.requiresAuth !== false && !isPublicPath && !await this.hasToken()) {
-      // Do NOT emit a GA4 error here — this is expected behavior when an
-      // unauthenticated user visits a protected page. Emitting it caused
-      // false-positive monitoring alerts (#9968, #9979, #9980, #9984).
-      throw new UnauthenticatedError()
-    }
-
-    // Check backend availability - waits for single health check on first load
-    const available = await checkBackendAvailability()
-    if (!available) {
-      throw new BackendUnavailableError()
-    }
-
-    const headers = { ...await this.getHeaders(), ...options?.headers }
-    const { controller, timeoutId } = this.createAbortController(options?.timeout ?? DEFAULT_TIMEOUT)
-    const signal = options?.signal ? AbortSignal.any([options.signal, controller.signal]) : controller.signal
-
-    try {
-      const response = await fetch(`${API_BASE}${path}`, {
-        method: 'GET',
-        headers,
-        signal,
-      })
-      clearTimeout(timeoutId)
-
-      if (!response.ok) {
-        // Handle 401 Unauthorized - token is invalid or expired.
-        // EXCLUDE per-feature endpoints whose 401 means a third-party token
-        // (e.g. GitHub OAuth) is missing/expired, NOT that the user's app
-        // session is dead. Logging the user out of the whole console because
-        // their GitHub token expired is a confusing dead end (e.g. clicking
-        // the "kubara" repo in the Mission Browser triggered a full logout).
-        // For these paths, surface the 401 to the caller so the feature can
-        // show its own auth prompt.
-        if (response.status === 401 && !path.startsWith('/api/github/')) {
-          handle401()
-          throw new UnauthorizedError()
-        }
-        if (response.status === 401) {
-          throw new UnauthorizedError()
-        }
-        if (response.status === 429) {
-          handle429(response)
-        }
-        const errorText = await safeReadTextOrEmpty(response, `[api] GET ${path} failed to read error response body`)
-        emitHttpError(String(response.status), errorText || '')
-        throw new Error(errorText || `API error: ${response.status}`)
-      }
-      markBackendSuccess()
-      this.checkTokenRefresh(response)
-      const data = await safeParseJsonOrNull<T>(response, `[api] GET ${path} failed to parse JSON response`)
-      return { data: data ?? ({} as T) }
-    } catch (err: unknown) {
-      clearTimeout(timeoutId)
-      if (err instanceof Error && err.name === 'AbortError') {
-        emitHttpError('timeout', `Request timeout after ${(options?.timeout ?? DEFAULT_TIMEOUT) / 1000}s`)
-        throw createErrorWithCause(`Request timeout after ${(options?.timeout ?? DEFAULT_TIMEOUT) / 1000}s: ${path}`, err)
-      }
-      if (err instanceof TypeError && err.message.includes('fetch')) {
-        markBackendFailure()
-        emitHttpError('network', err.message)
-      }
-      throw err
-    }
-  }
-
-  async post<T = unknown>(path: string, body?: unknown, options?: { timeout?: number; headers?: Record<string, string> }): Promise<{ data: T }> {
-    // Check backend availability
-    const available = await checkBackendAvailability()
-    if (!available) {
-      throw new BackendUnavailableError()
-    }
-
-    const { controller, timeoutId } = this.createAbortController(options?.timeout ?? DEFAULT_TIMEOUT)
-
-    try {
-      const response = await fetch(`${API_BASE}${path}`, {
-        method: 'POST',
-        headers: { ...await this.getHeaders(), ...options?.headers },
-        body: body ? JSON.stringify(body) : undefined,
-        signal: controller.signal,
-      })
-      clearTimeout(timeoutId)
-
-      if (!response.ok) {
-        // Handle 401 Unauthorized - token is invalid or expired.
-        // EXCLUDE per-feature endpoints (see GET handler comment) so a
-        // GitHub-OAuth-token expiry on /api/github/* doesn't log the user
-        // out of the entire console.
-        if (response.status === 401 && !path.startsWith('/api/github/')) {
-          handle401()
-          throw new UnauthorizedError()
-        }
-        if (response.status === 401) {
-          throw new UnauthorizedError()
-        }
-        if (response.status === 429) {
-          handle429(response)
-        }
-        const errorText = await safeReadTextOrEmpty(response, `[api] POST ${path} failed to read error response body`)
-        emitHttpError(String(response.status), errorText || '')
-        throw new Error(errorText || `API error: ${response.status}`)
-      }
-      markBackendSuccess()
-      this.checkTokenRefresh(response)
-      const data = await safeParseJsonOrNull<T>(response, `[api] POST ${path} failed to parse JSON response`)
-      return { data: data ?? ({} as T) }
-    } catch (err: unknown) {
-      clearTimeout(timeoutId)
-      if (err instanceof Error && err.name === 'AbortError') {
-        emitHttpError('timeout', `Request timeout after ${(options?.timeout ?? DEFAULT_TIMEOUT) / 1000}s`)
-        throw createErrorWithCause(`Request timeout after ${(options?.timeout ?? DEFAULT_TIMEOUT) / 1000}s: ${path}`, err)
-      }
-      if (err instanceof TypeError && err.message.includes('fetch')) {
-        markBackendFailure()
-        emitHttpError('network', err.message)
-      }
-      throw err
-    }
-  }
-
-  async patch<T = unknown>(path: string, body?: unknown, options?: { timeout?: number; headers?: Record<string, string> }): Promise<{ data: T }> {
-    // Check backend availability
-    const available = await checkBackendAvailability()
-    if (!available) {
-      throw new BackendUnavailableError()
-    }
-
-    const { controller, timeoutId } = this.createAbortController(options?.timeout ?? DEFAULT_TIMEOUT)
-
-    try {
-      const response = await fetch(`${API_BASE}${path}`, {
-        method: 'PATCH',
-        headers: { ...await this.getHeaders(), ...options?.headers },
-        body: body ? JSON.stringify(body) : undefined,
-        signal: controller.signal,
-      })
-      clearTimeout(timeoutId)
-
-      if (!response.ok) {
-        if (response.status === 401 && !path.startsWith('/api/github/')) {
-          handle401()
-          throw new UnauthorizedError()
-        }
-        if (response.status === 401) {
-          throw new UnauthorizedError()
-        }
-        if (response.status === 429) {
-          handle429(response)
-        }
-        const errorText = await safeReadTextOrEmpty(response, `[api] PATCH ${path} failed to read error response body`)
-        emitHttpError(String(response.status), errorText || '')
-        throw new Error(errorText || `API error: ${response.status}`)
-      }
-      markBackendSuccess()
-      this.checkTokenRefresh(response)
-      const data = await safeParseJsonOrNull<T>(response, `[api] PATCH ${path} failed to parse JSON response`)
-      return { data: data ?? ({} as T) }
-    } catch (err: unknown) {
-      clearTimeout(timeoutId)
-      if (err instanceof Error && err.name === 'AbortError') {
-        emitHttpError('timeout', `Request timeout after ${(options?.timeout ?? DEFAULT_TIMEOUT) / 1000}s`)
-        throw createErrorWithCause(`Request timeout after ${(options?.timeout ?? DEFAULT_TIMEOUT) / 1000}s: ${path}`, err)
-      }
-      if (err instanceof TypeError && err.message.includes('fetch')) {
-        markBackendFailure()
-        emitHttpError('network', err.message)
-      }
-      throw err
-    }
-  }
-
-  async put<T = unknown>(path: string, body?: unknown, options?: { timeout?: number }): Promise<{ data: T }> {
-    // Check backend availability
-    const available = await checkBackendAvailability()
-    if (!available) {
-      throw new BackendUnavailableError()
-    }
-
-    const { controller, timeoutId } = this.createAbortController(options?.timeout ?? DEFAULT_TIMEOUT)
-
-    try {
-      const response = await fetch(`${API_BASE}${path}`, {
-        method: 'PUT',
-        headers: await this.getHeaders(),
-        body: body ? JSON.stringify(body) : undefined,
-        signal: controller.signal,
-      })
-      clearTimeout(timeoutId)
-
-      if (!response.ok) {
-        // Handle 401 Unauthorized - token is invalid or expired.
-        // EXCLUDE per-feature endpoints (see GET handler comment) so a
-        // GitHub-OAuth-token expiry on /api/github/* doesn't log the user
-        // out of the entire console.
-        if (response.status === 401 && !path.startsWith('/api/github/')) {
-          handle401()
-          throw new UnauthorizedError()
-        }
-        if (response.status === 401) {
-          throw new UnauthorizedError()
-        }
-        if (response.status === 429) {
-          handle429(response)
-        }
-        const errorText = await safeReadTextOrEmpty(response, `[api] PUT ${path} failed to read error response body`)
-        emitHttpError(String(response.status), errorText || '')
-        throw new Error(errorText || `API error: ${response.status}`)
-      }
-      markBackendSuccess()
-      this.checkTokenRefresh(response)
-      const data = await safeParseJsonOrNull<T>(response, `[api] PUT ${path} failed to parse JSON response`)
-      return { data: data ?? ({} as T) }
-    } catch (err: unknown) {
-      clearTimeout(timeoutId)
-      if (err instanceof Error && err.name === 'AbortError') {
-        emitHttpError('timeout', `Request timeout after ${(options?.timeout ?? DEFAULT_TIMEOUT) / 1000}s`)
-        throw createErrorWithCause(`Request timeout after ${(options?.timeout ?? DEFAULT_TIMEOUT) / 1000}s: ${path}`, err)
-      }
-      if (err instanceof TypeError && err.message.includes('fetch')) {
-        markBackendFailure()
-        emitHttpError('network', err.message)
-      }
-      throw err
-    }
-  }
-
-  async delete(path: string, options?: { timeout?: number }): Promise<void> {
-    // Check backend availability
-    const available = await checkBackendAvailability()
-    if (!available) {
-      throw new BackendUnavailableError()
-    }
-
-    const { controller, timeoutId } = this.createAbortController(options?.timeout ?? DEFAULT_TIMEOUT)
-
-    try {
-      const response = await fetch(`${API_BASE}${path}`, {
-        method: 'DELETE',
-        headers: await this.getHeaders(),
-        signal: controller.signal,
-      })
-      clearTimeout(timeoutId)
-
-      if (!response.ok) {
-        // Handle 401 Unauthorized - token is invalid or expired.
-        // EXCLUDE per-feature endpoints (see GET handler comment) so a
-        // GitHub-OAuth-token expiry on /api/github/* doesn't log the user
-        // out of the entire console.
-        if (response.status === 401 && !path.startsWith('/api/github/')) {
-          handle401()
-          throw new UnauthorizedError()
-        }
-        if (response.status === 401) {
-          throw new UnauthorizedError()
-        }
-        if (response.status === 429) {
-          handle429(response)
-        }
-        const errorText = await safeReadTextOrEmpty(response, `[api] DELETE ${path} failed to read error response body`)
-        emitHttpError(String(response.status), errorText || '')
-        throw new Error(errorText || `API error: ${response.status}`)
-      }
-      markBackendSuccess()
-      this.checkTokenRefresh(response)
-    } catch (err: unknown) {
-      clearTimeout(timeoutId)
-      if (err instanceof Error && err.name === 'AbortError') {
-        emitHttpError('timeout', `Request timeout after ${(options?.timeout ?? DEFAULT_TIMEOUT) / 1000}s`)
-        throw createErrorWithCause(`Request timeout after ${(options?.timeout ?? DEFAULT_TIMEOUT) / 1000}s: ${path}`, err)
-      }
-      if (err instanceof TypeError && err.message.includes('fetch')) {
-        markBackendFailure()
-        emitHttpError('network', err.message)
-      }
-      throw err
-    }
-  }
-}
-
-function extractRequestPath(input: RequestInfo | URL): string {
-  const raw = typeof input === 'string'
-    ? input
-    : input instanceof URL
-      ? input.toString()
-      : input.url
-
-  try {
-    return new URL(raw, window.location.origin).pathname
-  } catch (error: unknown) {
-    reportAppError(error, {
-      context: '[api] failed to normalize request path',
-      level: 'warn',
-      fallbackMessage: 'request path normalization failed',
-    })
-    return raw
-  }
-}
-
-function shouldTreatAsBackendOutage(input: RequestInfo | URL, status: number): boolean {
-  if (!shouldMarkBackendUnavailable(status)) {
-    return false
-  }
-  const path = extractRequestPath(input)
-  return !BACKEND_OUTAGE_EXEMPT_PREFIXES.some(prefix => path.startsWith(prefix))
-}
-
-export const api = new ApiClient()
-
-/**
- * Drop-in replacement for `fetch()` that auto-injects the JWT Authorization
- * header from localStorage.  Use this for MCP endpoint calls that need auth
- * but return a raw Response (unlike `api.get()` which returns `{data}`).
- *
- * Existing callers only need to change `fetch(url, init)` -> `authFetch(url, init)`.
- */
-/**
- * Safely parse a Response as JSON, guarding against HTML responses.
- *
- * On Netlify, unmatched API routes fall through to the SPA catch-all which
- * returns index.html (200 OK, text/html). Calling `.json()` on that response
- * throws "Unexpected token '<'" (#9797). This helper checks the Content-Type
- * header first and throws a descriptive error instead of a cryptic parse error.
- *
- * Usage:
- *   const data = await safeJson<MyType>(response)
- */
-export async function safeJson<T = unknown>(response: Response): Promise<T> {
-  const contentType = response.headers.get('content-type') || ''
-  if (!contentType.includes('application/json')) {
-    throw new Error(
-      `Expected JSON response but received ${contentType || 'unknown content-type'} (status ${response.status})`,
-    )
-  }
-  return response.json() as Promise<T>
-}
-
-export async function authFetch(input: RequestInfo | URL, init?: RequestInit): Promise<Response> {
-  const token = await getStoredAuthToken()
-  const headers = new Headers(init?.headers)
-
-  if (token && token !== DEMO_TOKEN_VALUE && !headers.has('Authorization')) {
-    headers.set('Authorization', `Bearer ${token}`)
-  }
-  // #8830 — /api group RequireCSRF middleware rejects state-changing requests
-  // without this header; safeHTTPMethods pass through unconditionally, so
-  // setting it on every authFetch is correct and harmless for GET/HEAD.
-  if (!headers.has('X-Requested-With')) {
-    headers.set('X-Requested-With', 'XMLHttpRequest')
-  }
-
-  // Use caller-provided signal if present, otherwise apply default timeout
-  const signal = init?.signal ?? AbortSignal.timeout(FETCH_DEFAULT_TIMEOUT_MS)
-
-  try {
-    const response = await fetch(input, { ...init, headers, signal })
-    if (shouldTreatAsBackendOutage(input, response.status)) {
-      markBackendFailure(response.status)
-    } else {
-      markBackendSuccess(response.status)
-    }
-    const path = extractRequestPath(input)
-    if (response.status === 401 && !path.startsWith('/api/github/')) {
-      handle401()
-    }
-    return response
-  } catch (error) {
-    markBackendFailure()
-    throw error
-  }
-}
+export * from './api/index'
